Pricing Call and Put Options using Black-Scholes and Binomial Model
The task is to calculate the present value of call and put option using standard models i.e. Black-Scholes and Binomial Model. Day-count-convention (30/360 treats a month as 30 days and a year as 360 days) is used and options are priced mark-to-market  (the market data is assumed to be closing prices for that date).